recoverToTarget made extensible to allow multiple implementations
