--- server/src/main/java/org/opensearch/indices/recovery/remotestore/package-info.java
@@ -1,12 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-/**
- * Core classes responsible for handling all recovery of remote store enabled shards
- */
-package org.opensearch.indices.recovery.remotestore;
